+++ src/components/feature/CustomTooltip/CustomTooltipInfo.ts
@@ -4,16 +4,14 @@ import { getUuid } from '@/lib/lib_utils'
 
 export const version = '1.0.0'
 
-export const scopedId = getUuid('__i-popover__')
+export const scopedId = getUuid('__i-tooltip__')
 
 export type Placement = 'top' | 'top-start' | 'top-end' | 'bottom' | 'bottom-start' | 'bottom-end' | 'left' | 'left-start' | 'left-end' | 'right' | 'right-start' | 'right-end'
 export type Trigger = 'click' | 'focus' | 'hover' | 'contextmenu'
 
 export declare namespace Props {
   type Visible = boolean | null
-  type Width = number | string
-  type Title = string
-  type PopperStyle = string
+  type PopperClass = string
   type ShowArrow = boolean
   type Offset = number
 }
@@ -25,18 +23,6 @@ export const props = {
     default: null,
     description: '是否顯示'
   },
-  width: {
-    type: [Number, String] as PropType<Props.Width>,
-    required: false,
-    default: 150,
-    description: '寬度'
-  },
-  title: {
-    type: String as PropType<Props.Title>,
-    required: false,
-    default: '',
-    description: '標題'
-  },
   placement: {
     type: String as PropType<Placement>,
     required: false,
@@ -46,14 +32,14 @@ export const props = {
   trigger: {
     type: String as PropType<Trigger>,
     required: false,
-    default: 'click',
+    default: 'hover',
     description: '觸發方式'
   },
-  popperStyle: {
-    type: String as PropType<Props.PopperStyle>,
+  popperClass: {
+    type: String as PropType<Props.PopperClass>,
     required: false,
     default: '',
-    description: '樣式'
+    description: 'class'
   },
   showArrow: {
     type: Boolean as PropType<Props.ShowArrow>,