+++ src/components/feature/CustomTree/CustomTree.vue
@@ -1,65 +1,16 @@
 <script setup lang="ts">
-import { type PropType, useSlots, ref } from 'vue'
+import { useSlots, ref } from 'vue'
 import { ElTree } from 'element-plus'
 import type Node from 'element-plus/es/components/tree/src/model/node'
 
-import { getUuid } from '@/lib/lib_utils'
+import type { CheckNode } from './CustomTreeInfo'
+import {
+  version,
+  scopedId,
+  props as treeProps
+} from './CustomTreeInfo'
 
-export interface TreeNodeData {
-  [key: string]: any
-}
-
-export interface TreeOptionProps {
-  children?: string
-  label?: string | ((data: TreeNodeData, node: Node) => string)
-  disabled?: string | ((data: TreeNodeData, node: Node) => string)
-  isLeaf?: string | ((data: TreeNodeData, node: Node) => boolean)
-  class?: (data: TreeNodeData, node: Node) => string | {
-      [key: string]: boolean
-  } | string
-}
-
-const props = defineProps({
-  data: {
-    type: Array as PropType<any[]>,
-    required: false,
-    default: () => {
-      return []
-    },
-    description: '樹結構資料'
-  },
-  props: {
-    type: Object as PropType<TreeOptionProps | any>,
-    required: false,
-    default: () => {
-      return {}
-    },
-    description: '參數'
-  },
-  nodeKey: {
-    type: String as PropType<string>,
-    required: false,
-    description: '節點資料唯一值'
-  },
-  highlightCurrent: {
-    type: Boolean as PropType<boolean>,
-    required: false,
-    default: false,
-    description: '是否有背景顏色 在節點被選中'
-  },
-  defaultExpandAll: {
-    type: Boolean as PropType<boolean>,
-    required: false,
-    default: false,
-    description: '是否展開所有節點'
-  },
-  showCheckbox: {
-    type: Boolean as PropType<boolean>,
-    required: false,
-    default: false,
-    description: '是否使用checkbox'
-  }
-})
+const props = defineProps(treeProps)
 
 const emit = defineEmits([
   'node-click',
@@ -75,12 +26,6 @@ const onCheckChange = ($data1: any, $data2: any, $data3: any) => {
 }
 const onCheck = (nodeDate: any, checkedData: any) => {
   emit('check', nodeDate, checkedData)
-}
-
-// slot
-const slots = useSlots()
-const hasSlot = (prop: string): boolean => {
-  return !!slots[prop]
 }
 
 const elTreeRef = ref<InstanceType<typeof ElTree>>()
@@ -106,7 +51,6 @@ const resetChecked = () => {
   elTreeRef.value!.setCheckedKeys([], false)
 }
 
-type CheckNode = Node | string | number
 const setChecked = (checkNode: CheckNode, checked: boolean, deep: boolean) => {
   elTreeRef.value!.setChecked(checkNode, checked ?? false, deep ?? false)
 }
@@ -120,12 +64,15 @@ defineExpose({
   setChecked
 })
 
-const scopedId = getUuid('__i-tree__')
+const slots = useSlots()
+const hasSlot = (prop: string): boolean => {
+  return !!slots[prop]
+}
 
 </script>
 
 <template>
-  <div class="__tree-wrapper" :class="scopedId">
+  <div :class="`CustomTree_${version} ${scopedId}`" class="__tree-wrapper">
     <ElTree
       ref="elTreeRef"
       :data="props.data"